requirements.txt
Babel==1.3
Django==1.7.7
Jinja2==2.7.3
MarkupSafe==0.23
MySQL-python==1.2.5
PyJWT==0.4.3
PyYAML==3.11
amqp==1.4.6
anyjson==0.3.3
argparse==1.2.1
backports.ssl-match-hostname==3.4.0.2
billiard==3.3.0.19
celery==3.1.17
certifi==14.05.14
dj-database-url==0.3.0
django-celery==3.1.16
django-cors-headers==1.0.0
django-filter==0.9.2
django-grip==1.6.1
django-sse==0.4.1
django-sse-wrapper==0.0.3
django-websocket-redis==0.4.4
ecdsa==0.13
elasticsearch==1.4.0
eventlet==0.17.3
futures==2.2.0
gevent==1.0.1
gevent-websocket==0.9.3
google-api-python-client==1.2
greenlet==0.4.5
gripcontrol==2.3.2
gunicorn==19.3.0
httplib2==0.9
jsonfield==1.0.3
kombu==3.0.24
paramiko==1.15.2
pubcontrol==2.1.1
pycrypto==2.6.1
pyelasticsearch==1.2.1
python-dateutil==2.4.2
pytz==2015.2
redis==2.10.3
requests==2.6.0
simplejson==3.6.5
six==1.9.0
sockjs-tornado==1.0.1
sse==1.2
urllib3==1.10.2
wsgiref==0.1.2